+++ src/enums.py
@@ -6,10 +6,10 @@
 
 from enum import Enum
 
-class Model(Enum):
-    LR=1
-    CNN=2
-    SVM=3
+class Model():
+    LR='logistic_regression'
+    CNN='cnn'
+    SVM='svm'
 
 class Dataset():
     BENEVOLENT='benevolent'
@@ -19,13 +19,13 @@ class Dataset():
     SCALES='scales'
     
 class Domain():
-    BHO='bho'
-    C='c'
-    BHOCS='bhocs'
-    S='s'
-    BHOM='bhom'
-    CM='cm'
-    BHOCSM='bhocsm'
+    BHO={'modified': False, 'dataset': [Dataset.BENEVOLENT, Dataset.HOSTILE, Dataset.OTHER]}
+    C={'modified': False, 'dataset': [Dataset.CALLME]}
+    BHOCS={'modified': False, 'dataset': [Dataset.BENEVOLENT, Dataset.HOSTILE, Dataset.OTHER, Dataset.CALLME, Dataset.SCALES]}
+    S={'modified': False, 'dataset': [Dataset.SCALES]}
+    BHOM={'modified': True, 'dataset': [Dataset.BENEVOLENT, Dataset.HOSTILE, Dataset.OTHER]}
+    CM={'modified': True, 'dataset': [Dataset.CALLME]}
+    BHOCSM={'modified': True, 'dataset': [Dataset.BENEVOLENT, Dataset.HOSTILE, Dataset.OTHER, Dataset.CALLME, Dataset.SCALES]}
     
 class Label():
     NONSEXIST=int('0')
@@ -44,3 +44,12 @@ class Parameter():
             NGRAM_RANGE='ngram_range' # (2,2)
     class TypeDependency():
             MODEL_PATH='model_path'
+            NGRAM_RANGE='ngram_range' # (2,2)
+   
+            
+class DataColumn():
+    ID='_id'
+    ADVERSARIAL='of_id'
+    DATASET='dataset'
+    TEXT='text'
+    LABEL='sexist'