+++ src/item_selector.py
@@ -6,14 +6,17 @@
 #sklearn
 from sklearn.base import BaseEstimator, TransformerMixin
 
-class ItemSelector(BaseEstimator, TransformerMixin):
+#other
+import inspect
+from collections import defaultdict
 
-    def __init__(self, key):
+class ItemSelector(BaseEstimator):
+    
+    def __init__(self, key=''):
         self.key = key
 
     def fit(self, x, y=None):
-        return self
+        return self #check for ngrams
 
     def transform(self, data_dict):
         return data_dict[self.key]
-